ref: moves encoder files to pkg folder
