+++ api/middlewares/auth.go
@@ -3,13 +3,14 @@ package middlewares
 import (
 	"net/http"
 	"pet-dex-backend/v2/infra/config"
+	"pet-dex-backend/v2/pkg/encoder"
 	"strings"
 	"time"
 )
 
 func AuthMiddleware(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		encoder := NewEncoderAdapter(config.GetEnvConfig().JWT_SECRET)
+		encoder := encoder.NewEncoderAdapter(config.GetEnvConfig().JWT_SECRET)
 
 		authHeader := r.Header.Get("Authorization")
 		if authHeader == "" {
