+++ pkg/encoder/encoder.go
@@ -1,4 +1,4 @@
-package middlewares
+package encoder
 
 import (
 	"pet-dex-backend/v2/interfaces"
